+++ src/PJ/pjhigh/flipdr.c
@@ -87,7 +87,7 @@ done:
 }
 static Errcode fliif_save_frames(Image_file *ifile, 
 						   		 Rcel *screen, 
-						   		 int num_frames, 
+								 ULONG num_frames,
 						   		 Errcode (*seek_frame)(int ix,void *seek_data),
 						   		 void *seek_data,
 						   		 Rcel *work_screen )
